@@ -1,25 +1,25 @@
 name: cpx_research_sdk_flutter
 description: CPX Research SDK Flutter
-version: 0.4.3
+version: 0.4.4
 homepage: https://www.makeopinion.com
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.18.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.4
-  webview_flutter: ^3.0.4
-  json_annotation: ^4.5.0
+  http: ^1.1.0
+  webview_flutter: ^4.2.2
+  json_annotation: ^4.7.0
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   build_runner: ^2.1.8
-  json_serializable: ^6.1.5
+  json_serializable: ^6.6.1
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
