@@ -23,15 +23,17 @@ class _BrowserViewState extends State<BrowserView> {
   Controller controller = Controller.controller;
   bool isLoading = true;
   WebViewController? _controller;
-  late List pages;
+  late List<Uri> pages;
   late BrowserTab activeTab;
   bool isAlertDisplayed = false;
 
   /// [loadURL] loads the url in the webview
   void loadURL(int index) {
     if (_controller != null) {
-      _controller!.loadUrl(pages[index]);
-      CPXLogger.log("Load url: " + pages[index]);
+      _controller!.loadRequest(pages[index]);
+      CPXLogger.log("Load url: ${pages[index]}");
+    } else {
+      CPXLogger.log("Failed to load url: ${pages[index]} | WebViewController is null");
     }
   }
 
@@ -39,11 +41,35 @@ class _BrowserViewState extends State<BrowserView> {
   void initState() {
     activeTab = widget.currentTab;
     pages = [
-      NetworkService().getHomeURL().toString(),
-      NetworkService().getSettingsURL().toString(),
-      NetworkService().getHelpURL().toString(),
+      NetworkService().getHomeURL(),
+      NetworkService().getHelpURL(),
     ];
+    initWebView();
     super.initState();
+  }
+
+  initWebView() {
+    _controller = WebViewController()
+      ..setJavaScriptMode(JavaScriptMode.unrestricted)
+      ..setBackgroundColor(const Color(0x00000000))
+      ..setNavigationDelegate(
+        NavigationDelegate(
+          onPageStarted: (start) => setState(() => isLoading = true),
+          onPageFinished: (finish) => setState(() => isLoading = false),
+          onWebResourceError: (error) {
+            HapticFeedback.selectionClick();
+            if ((error.errorCode == -1 && Platform.isAndroid) ||
+                (error.errorCode == -1022 && Platform.isIOS)) {
+              setState(() => isAlertDisplayed = true);
+            }
+            CPXLogger.log(
+                "Browser error: " + error.errorCode.toString() + " | " + error.description);
+            NetworkService().onWebViewError(
+                error.errorCode.toString(), error.description, error.url ?? "no url");
+          },
+        ),
+      )
+      ..loadRequest(pages[activeTab == BrowserTab.settings ? 1 : 0]);
   }
 
   @override
@@ -59,17 +85,16 @@ class _BrowserViewState extends State<BrowserView> {
                 Container(
                   margin: const EdgeInsets.symmetric(horizontal: 1),
                   decoration: new BoxDecoration(
-                    color: activeTab == BrowserTab.help
-                        ? controller.config.accentColor
-                        : Colors.white,
-                    borderRadius:
-                        BorderRadius.vertical(top: Radius.circular(10.0)),
+                    color:
+                        activeTab == BrowserTab.help ? controller.config.accentColor : Colors.white,
+                    borderRadius: BorderRadius.vertical(top: Radius.circular(10.0)),
                   ),
                   child: IconButton(
                     icon: Icon(Icons.help_outline),
                     onPressed: () {
                       HapticFeedback.selectionClick();
-                      loadURL(2);
+                      loadURL(1);
+                      isAlertDisplayed = false;
                       activeTab = BrowserTab.help;
                     },
                     color: Colors.black,
@@ -78,36 +103,16 @@ class _BrowserViewState extends State<BrowserView> {
                 Container(
                   margin: const EdgeInsets.symmetric(horizontal: 1),
                   decoration: new BoxDecoration(
-                    color: activeTab == BrowserTab.settings
-                        ? controller.config.accentColor
-                        : Colors.white,
-                    borderRadius:
-                        BorderRadius.vertical(top: Radius.circular(10.0)),
-                  ),
-                  child: IconButton(
-                    icon: Icon(Icons.settings_outlined),
-                    onPressed: () {
-                      HapticFeedback.selectionClick();
-                      loadURL(1);
-                      activeTab = BrowserTab.settings;
-                    },
-                    color: Colors.black,
-                  ),
-                ),
-                Container(
-                  margin: const EdgeInsets.symmetric(horizontal: 1),
-                  decoration: new BoxDecoration(
-                    color: activeTab == BrowserTab.home
-                        ? controller.config.accentColor
-                        : Colors.white,
-                    borderRadius:
-                        BorderRadius.vertical(top: Radius.circular(10.0)),
+                    color:
+                        activeTab == BrowserTab.home ? controller.config.accentColor : Colors.white,
+                    borderRadius: BorderRadius.vertical(top: Radius.circular(10.0)),
                   ),
                   child: IconButton(
                     icon: Icon(Icons.home_outlined),
                     onPressed: () {
                       HapticFeedback.selectionClick();
                       loadURL(0);
+                      isAlertDisplayed = false;
                       activeTab = BrowserTab.home;
                     },
                     color: Colors.black,
@@ -117,14 +122,14 @@ class _BrowserViewState extends State<BrowserView> {
                   margin: const EdgeInsets.symmetric(horizontal: 1),
                   decoration: new BoxDecoration(
                     color: Colors.white,
-                    borderRadius:
-                        BorderRadius.vertical(top: Radius.circular(10.0)),
+                    borderRadius: BorderRadius.vertical(top: Radius.circular(10.0)),
                   ),
                   child: IconButton(
                     icon: Icon(Icons.close),
                     onPressed: () {
                       HapticFeedback.selectionClick();
                       CPXLogger.log("Close CPX Browser");
+                      isAlertDisplayed = false;
                       Controller.controller.showWidgets();
                     },
                     color: Colors.black,
@@ -133,39 +138,20 @@ class _BrowserViewState extends State<BrowserView> {
               ],
             ),
             Expanded(
-              child: Stack(
-                children: [
-                  WebView(
-                    onWebViewCreated: (WebViewController c) => _controller = c,
-                    javascriptMode: JavascriptMode.unrestricted,
-                    initialUrl: pages[activeTab == BrowserTab.settings ? 1 : 0],
-                    onPageFinished: (finish) {
-                      setState(() => isLoading = false);
-                    },
-                    onPageStarted: (start) {
-                      setState(() => isLoading = true);
-                    },
-                    onWebResourceError: (error) {
-                      HapticFeedback.selectionClick();
-                      setState(() => isAlertDisplayed = true);
-                      CPXLogger.log("Browser error: " +
-                          error.errorCode.toString() +
-                          " | " +
-                          error.description);
-                      NetworkService().onWebViewError(
-                          error.errorCode.toString(),
-                          error.description,
-                          error.failingUrl ?? "no url");
-                    },
-                  ),
-                  if (isLoading)
-                    LinearProgressIndicator(
-                      valueColor: new AlwaysStoppedAnimation<Color>(
-                          controller.config.accentColor),
-                      backgroundColor: Colors.white,
-                    ),
-                  if (isAlertDisplayed) ErrorAlertDialog(),
-                ],
+              child: ColoredBox(
+                color: Colors.white,
+                child: Stack(
+                  children: [
+                    WebViewWidget(controller: _controller!),
+                    if (isLoading)
+                      LinearProgressIndicator(
+                        valueColor:
+                            new AlwaysStoppedAnimation<Color>(controller.config.accentColor),
+                        backgroundColor: Colors.white,
+                      ),
+                    if (isAlertDisplayed) showReloadOnError(),
+                  ],
+                ),
               ),
             ),
           ],
@@ -174,34 +160,89 @@ class _BrowserViewState extends State<BrowserView> {
     );
   }
 
-  Container ErrorAlertDialog() {
-    Widget textButton = TextButton(
-      child: Text(
-        "OK",
-        style: TextStyle(color: controller.config.accentColor),
-      ),
-      onPressed: () {
-        HapticFeedback.selectionClick();
-        setState(() => isAlertDisplayed = false);
-        CPXLogger.log("Close CPX Browser");
-        Controller.controller.showWidgets();
-      },
-    );
-    return Container(
-      color: Colors.black87,
-      child: Platform.isIOS
-          ? CupertinoAlertDialog(
-              title: Text("Browser Error"),
-              content:
-                  Text("An error occurred, while using the survey browser"),
-              actions: [textButton],
-            )
-          : AlertDialog(
-              title: Text("Browser Error"),
-              content:
-                  Text("An error occurred, while using the survey browser"),
-              actions: [textButton],
+  Widget showReloadOnError() => Center(
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+          children: [
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(10.0),
+                ),
+                backgroundColor: controller.config.accentColor,
+                foregroundColor: Colors.black,
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Icon(Icons.arrow_back_ios_new_rounded, size: 20),
+                  SizedBox(width: 4),
+                  Text(
+                    "Go back",
+                    style: TextStyle(fontSize: 20),
+                  ),
+                ],
+              ),
+              onPressed: () {
+                HapticFeedback.selectionClick();
+                setState(() => isAlertDisplayed = false);
+                _controller!.goBack();
+              },
             ),
-    );
-  }
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(10.0),
+                ),
+                backgroundColor: controller.config.accentColor,
+                foregroundColor: Colors.black,
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Icon(Icons.refresh, size: 20),
+                  SizedBox(width: 4),
+                  Text(
+                    "Reload",
+                    style: TextStyle(fontSize: 20),
+                  ),
+                ],
+              ),
+              onPressed: () {
+                HapticFeedback.selectionClick();
+                setState(() => isAlertDisplayed = false);
+                _controller!.reload();
+              },
+            ),
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(10.0),
+                ),
+                backgroundColor: controller.config.accentColor,
+                foregroundColor: Colors.black,
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Text(
+                    "Opinions",
+                    style: TextStyle(fontSize: 20),
+                  ),
+                  SizedBox(width: 4),
+                  Icon(Icons.arrow_forward_ios_rounded, size: 20),
+                ],
+              ),
+              onPressed: () {
+                HapticFeedback.selectionClick();
+                setState(() => isAlertDisplayed = false);
+                loadURL(0);
+              },
+            ),
+          ],
+        ),
+      );
 }
